+++ app/processor.py
@@ -1,6 +1,8 @@
 from app.helpers.request_helper import get_sequence_no
 import json
 from app import settings
+import re
+from app.helpers.exceptions import BadMessageError, RetryableError
 
 
 class CTPProcessor(object):
@@ -9,10 +11,28 @@ class CTPProcessor(object):
         self.logger = logger
         self.survey = survey
         self.tx_id = None
-        self.setup_logger()
+        self._setup_logger()
         self.ftp = ftpconn
 
-    def setup_logger(self):
+    def process(self):
+        sequence_no = get_sequence_no()
+        if sequence_no is None:
+            raise RetryableError("Failed to get sequence number")
+
+        filename = '{}.json'.format(sequence_no)
+
+        if self.survey is None:
+            raise BadMessageError("No survey data")
+
+        data = json.dumps(self.survey)
+
+        # Attempt to deliver the real file and send a .completed after it
+        self._deliver_file(filename, data)
+        self._deliver_file(self._get_completed_filename(filename), "")
+
+        return
+
+    def _setup_logger(self):
         if self.survey:
             if 'metadata' in self.survey:
                 metadata = self.survey['metadata']
@@ -22,28 +42,22 @@ class CTPProcessor(object):
                 self.tx_id = self.survey['tx_id']
                 self.logger = self.logger.bind(tx_id=self.tx_id)
 
-    def deliver_file(self, filename, data):
-        folder = self.get_ftp_folder(self.survey)
-        return self.ftp.deliver_binary(folder, filename, data.encode('utf-8'))
+    def _deliver_file(self, filename, data):
+        self.logger.debug("Attempting to deliver to ftp", filename=filename)
+        folder = self._get_ftp_folder(self.survey)
 
-    def process(self):
-        filename = '{}.json'.format(get_sequence_no())
-        data = json.dumps(self.survey)
+        try:
+            self.ftp.deliver_binary(folder, filename, data.encode('utf-8'))
+        except (RuntimeError, Exception) as e:
+            self.logger.error("Failed to deliver file to ftp", filename=filename, exception=e)
+            raise RetryableError("Failed to deliver file to ftp")
 
-        if data is None:
-            return False
+        self.logger.info("Delivered file to ftp", filename=filename)
 
-        # Attempt to deliver the real file and send
-        # a .completed after it
-        self.deliver_file(filename, data)
+    def _get_completed_filename(self, filename):
+        return re.compile('.json$').sub('.completed', filename)
 
-        completed_filename = filename + ".completed"
-        self.logger.info("Sending 'completed file'", filename=completed_filename)
-        self.deliver_file(completed_filename, "")
-
-        return
-
-    def get_ftp_folder(self, survey):
+    def _get_ftp_folder(self, survey):
         if 'heartbeat' in survey and survey['heartbeat'] is True:
             return settings.FTP_HEARTBEAT_FOLDER
         else: