+++ CHANGELOG.md
@@ -1,2 +1,13 @@
 ### Unreleased
+  - Add environment variables to README
+  
+### 2.0.0 2017-03-15
+  - Log version number on startup
+  - Fix handling of None responses in remote call
+  - Add SFTP support
+  - Change `nack for retry` to `nack` for SDX logging
+  - Change `status_code` to `status` for SDX logging
+  - Change logging messages to add the service called or returned from
+
+### 1.0.0 2017-02-16
   - Initial release